Add more compact roads data
This road data can be simplified down to 1MB or so with topojson, instead of 17MB.
It's from https://www.naturalearthdata.com/downloads/10m-cultural-vectors/roads/

The new script filters out needless features, limits to the USA, and keeps things like
road name, and what type of highway it is. We can further reduce data size by removing
the road name and scale-rank if needed.